Restructure folders
Restructure folders,
Remove extra codes
