+++ .idea/workspace.xml
@@ -62,7 +62,7 @@
   <component name="ProjectLevelVcsManager" settingsEditedManually="true" />
   <component name="PropertiesComponent">
     <property name="WebServerToolWindowFactoryState" value="false" />
-    <property name="last_opened_file_path" value="$PROJECT_DIR$/store" />
+    <property name="last_opened_file_path" value="$PROJECT_DIR$/store/tests" />
     <property name="settings.editor.selected.configurable" value="com.jetbrains.python.configuration.PyActiveSdkModuleConfigurable" />
   </component>
   <component name="RecentsManager">
@@ -162,7 +162,7 @@
       <workItem from="1622786549694" duration="5267000" />
       <workItem from="1622795814245" duration="313000" />
       <workItem from="1622800943853" duration="7338000" />
-      <workItem from="1622832831375" duration="33312000" />
+      <workItem from="1622832831375" duration="35132000" />
     </task>
     <servers />
   </component>
@@ -186,7 +186,7 @@
     <breakpoint-manager>
       <breakpoints>
         <line-breakpoint enabled="true" suspend="THREAD" type="python-line">
-          <url>file://$PROJECT_DIR$/store/_tests/rest_full_test.py</url>
+          <url>file://$PROJECT_DIR$/store/tests/rest_full_test.py</url>
           <line>16</line>
           <option name="timeStamp" value="22" />
         </line-breakpoint>
@@ -209,6 +209,6 @@
     </watches-manager>
   </component>
   <component name="com.intellij.coverage.CoverageDataManagerImpl">
-    <SUITE FILE_PATH="coverage/django_devzillas_store$test.coverage" NAME="test Coverage Results" MODIFIED="1622922946451" SOURCE_PROVIDER="com.intellij.coverage.DefaultCoverageFileProvider" RUNNER="coverage.py" COVERAGE_BY_TEST_ENABLED="true" COVERAGE_TRACING_ENABLED="false" WORKING_DIRECTORY="" />
+    <SUITE FILE_PATH="coverage/django_devzillas_store$test.coverage" NAME="test Coverage Results" MODIFIED="1622925887183" SOURCE_PROVIDER="com.intellij.coverage.DefaultCoverageFileProvider" RUNNER="coverage.py" COVERAGE_BY_TEST_ENABLED="true" COVERAGE_TRACING_ENABLED="false" WORKING_DIRECTORY="" />
   </component>
 </project>
