+++ devzilas_store/urls.py
@@ -1,13 +1,13 @@
 """devzilas_store URL Configuration
 
-The `urlpatterns` list routes URLs to _views. For more information please see:
+The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/3.0/topics/http/urls/
 Examples:
-Function _views
-    1. Add an import:  from my_app import _views
-    2. Add a URL to urlpatterns:  path('', _views.home, name='home')
-Class-based _views
-    1. Add an import:  from other_app._views import Home
+Function views
+    1. Add an import:  from my_app import views
+    2. Add a URL to urlpatterns:  path('', views.home, name='home')
+Class-based views
+    1. Add an import:  from other_app.views import Home
     2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
 Including another URLconf
     1. Import the include() function: from django.urls import include, path
