+++ store/admin.py
@@ -1,5 +1,6 @@
 from django.contrib import admin
-from .models import Department, Feature
+from .models. department import Department
+from .models.feature import Feature
 
 admin.site.register(Department)
 admin.site.register(Feature)
