--- store/models.py
@@ -1,5 +1,0 @@
-from ._models.department import Department
-from ._models.feature import Feature
-
-__all__ = ['Department', 'Feature']
-
