--- store/tests.py
@@ -1,4 +1,0 @@
-from ._tests.department import DepartmentTests
-from ._tests.feature import FeatureTests
-
-__all__ = ['DepartmentTests', 'FeatureTests']
