+++ store/tests/__init__.py
@@ -1,0 +1,2 @@
+from .department import DepartmentTests
+from .feature import FeatureTests
