--- store/views.py
@@ -1,5 +1,0 @@
-from ._views.department import AddDepartment, DepartmentDetail, DepartmentList
-from ._views.feature import AddFeature, FeatureDetail, FeatureList
-
-__all__ =['AddDepartment', 'DepartmentDetail', 'DepartmentList',
-          'AddFeature', 'FeatureDetail', 'FeatureList']
